Event indexing refactoring.
